fixed, use sonar and quickstart

+++ homework/Vasili_Zasinets/ArrayServiceTask/src/main/java/com/tutrit/java/quickstart/homework/arrayservice/ArrayService.java
@@ -36,12 +36,13 @@ public class ArrayService {
     public double avg(int[] array) {
         if (verification(array) != 0) {
             int sumOfArray = 0;
-            int average;
+            double average;
             for (int j : array) {
-                sumOfArray += j;}
-            average = sumOfArray / (array.length);
-            return average;}
-        else return 0;
+                sumOfArray += j;
+            }
+            average = (double) sumOfArray / (array.length);
+            return average;
+        } else return 0;
     }
 
     public int verification(int[] array) {
